+++ sale_discount_total/models/discount_approval.py
@@ -20,7 +20,7 @@
 #
 #############################################################################
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class sale_discount(models.Model):