+++ sale_discount_total/doc/RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-## Module <inventory_barcode_scanning>
+## Module <sale_discount_total>
 
 #### 08.10.2020
 #### Version 14.0.1.0.0
@@ -10,5 +10,8 @@ Initial commit for Sale Discount On Total Amount
 #### FIX
 discount roundoff
 
-
+#### 02.02.2024
+#### Version 14.0.1.2.0
+#### FIX
+discount roundoff in website